Move Config{,Section} to rhsm.config
This was done to enable use of the somewhat generic Config and ConfigSection classes
in other tools (for example Virt-who) that could benefit from a derivative of such a
structure.
